requirements.txt
tinydb==3.1.0
python-telegram-bot==3.1.2
requests==2.9.1
python-chess==0.13.1
Pillow==3.1.0
